add some comments and refactor the FS loader creation into a separate method
