+++ scripts/content-repo-preview/builders/io-sites.ts
@@ -6,6 +6,9 @@ import { execFileSync } from 'child_process'
 const cwd = process.cwd()
 const globalCSSFile = path.join(cwd, 'src', 'pages', 'style.css')
 
+/**
+ * Handles building deploy previews for content repositories within the context of a product's .io site.
+ */
 export function IoPreviewBuilder(product) {
 	return {
 		async prebuild() {
